Executando abertura de janelas sem títulos. Usando o Windows handles para buscar o id da página e armazenar no Array e alternar para página principal.
